Make e1000 work in VirtualBox
This patch makes the e1000 driver work in VirtualBox. I'm working
towards getting the driver working in VMWare as well, but VMWare
emulates a different chip (82545 rather than 82540). Those fixes and
adjustments will come in a different patch, at a later date.

The biggest change here is that the driver now supports the microwire
interface for reading from the EEPROM as well as the EERD.

Signed-off-by: Matthew Iselin <matthew@theiselins.net>
Signed-off-by: Kevin Wolf <kevin@tyndur.org>