cdi.h: Add cdi_provide_device_internal_drv()
cdi_provide_device() hasn't been implemented so far because it raises
questions of inter-driver communication in microkernel environments that
aren't trivial to answer.

This patch adds an interface to provide a device that is guaranteed
to be handled by a CDI driver in the same binary, i.e. when passing
pointers works.

This is not meant as a general solution, but only to be used as a means
for the internal organisation of complex drivers where some kind of an
internal bus exists that can neither be exposed by a different driver
nor can its children be handled by a different driver.

Signed-off-by: Kevin Wolf <kevin@tyndur.org>