iso9660: Fix resource flags
It's not a very good idea to flag every resource as readable, browsable
and symlink-readable at the same time. This fixes the flags to actually
match the assigned class.

The iso9660 driver knows symlinks as it reads them from a Rockridge
image, however it doesn't implement the .read_link callback, so using
the respective flag doesn't make sense. Treat symlinks as regular files
for now.

Signed-off-by: Kevin Wolf <kevin@tyndur.org>