doxyfile.in: Define aliases for French, Japanese and Saxonian
Our dear Max likes to be silly with multilingual comments. Fine with me,
but make sure that this stuff doesn't end up in English or German
documentation.

This patch doesn't change the Makefile to actually produce documentation
in any of the new languages as it would be utterly useless in its
incomplete state.

Signed-off-by: Kevin Wolf <kevin@tyndur.org>