e1000: Fix EERD reading for e1000e
The EERD reading function missed that the offset for the EEPROM address
differs between e1000 and e1000e, too. Instead of adding yet another
parameter to do_read_eerd() that has to be passed by model-specific
wrappers, put the parameter directly in the model struct and take them
from there.

Signed-off-by: Kevin Wolf <kevin@tyndur.org>