e1000: Acknowledge interrupts explicitly by ICR write
The old e1000 models used to clear the interrupt flags in ICR on read,
but e1000e doesn't do this any more. Just writing the bits back is fine
for both, so let's do just that.

Clearing the bits must be done before actually processing the interrupt
in order to avoid races with new events.

Signed-off-by: Kevin Wolf <kevin@tyndur.org>